// model/orderDiscountLinesRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* List of discounts that apply to the order.
*/
export class OrderDiscountLinesRequest {
    /**
    * The amount to be deducted from the total sum of all payments, in cents.
    */
    'amount': number;
    /**
    * Discount code.
    */
    'code': string;
    /**
    * It can be \'loyalty\', \'campaign\', \'coupon\' o \'sign\'
    */
    'type': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return OrderDiscountLinesRequest.attributeTypeMap;
    }
}